Added graphic image ZPL2 command (^GF) with ascii encoding (GRF format)
The ascii encoding is reverse engineered from the little info that
i found on this topic on page 149 and 150 on hte ZPL programming guide,
in the section documenting the ~DG Download Graphics command.

The ascii encoding is not efficient, it increases the image site by a
factor of 2.  There is an optional compression scheme which is not
implemented yet.